refactor: Update UiSizes class with MediaQuery size and properties
This commit updates the UiSizes class in uisizes.dart to use the MediaQuery class for obtaining the size, pixel ratio, orientation, and aspect ratio of the app's screen. The previous code was using the deprecated MediaQuery.of(context) method, which has been replaced with the MediaQuery.sizeOf(context), MediaQuery.devicePixelRatioOf(context), MediaQuery.orientationOf(context), and MediaQuery.sizeOf(context).aspectRatio methods. This refactoring improves the code's compatibility with the latest Flutter version and ensures accurate measurement of the app's UI elements.