refactor: Update main.dart to use DevicePreview for previewing app on different devices
This commit updates the main.dart file to import the 'device_preview' package and enable the DevicePreview widget. The DevicePreview widget is wrapped around the existing MainApp widget and is only enabled in non-release mode. This enhancement allows developers to preview the app on different devices during development, improving the testing and debugging process.